Update compile_firmwares.yml workflow
This commit updates the compile_firmwares.yml workflow file. It removes the working directory for the compile_one_button_sketch job and adds the working directory for the steps that compile, list files, create firmware, and upload firmware for the OneButton sketch. It also adds the working directory for the compile_m5cardputer_user_demo job and the steps that compile, list files, and upload firmware for the M5Cardputer-UserDemo.